Switched the main code into the app folder.  Need to set environment variable FLASK_ENV to 'development' to enter DEBUG mode.
